+++ lib/models/catalog.dart
@@ -6,7 +6,6 @@ import '@models.dart';
 import 'dart:convert';
 
 Catalog catalogFromJson(String str) => Catalog.fromJson(json.decode(str));
-
 String catalogToJson(Catalog data) => json.encode(data.toJson());
 
 class Catalog {