+++ lib/router.dart
@@ -10,6 +10,9 @@ Route<dynamic> generateRoute(RouteSettings settings) {
     case HomeRoute:
       return MaterialPageRoute(
           builder: (context) => HomeForm(message: settings.arguments));
+    case UserRoute:
+      return MaterialPageRoute(
+          builder: (context) => UserForm(settings.arguments));
     case LoginRoute:
       return MaterialPageRoute(
           builder: (context) => LoginForm(settings.arguments));