+++ test/blocs/auth_bloc_test.dart
@@ -2,7 +2,7 @@ import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:dio/dio.dart';
-import 'package:hotel/blocs/@bloc.dart';
+import 'package:hotel/blocs/@blocs.dart';
 import 'package:hotel/services/repos.dart';
 import '../data.dart';
 
@@ -57,7 +57,8 @@ void main() {
         when(mockReposRepository.getConnected()).thenAnswer((_) async => true);
         when(mockReposRepository.getAuthenticate())
             .thenAnswer((_) async => authenticate);
-        authenticate.apiKey = 'dummyKey';
+        when(mockReposRepository.checkApikey(authenticate.apiKey))
+            .thenAnswer((_) async => true);
         bloc.add(LoadAuth());
       },
       expect: <AuthState>[