+++ test/blocs/changePw_bloc_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
-import 'package:hotel/blocs/@bloc.dart';
+import 'package:hotel/blocs/@blocs.dart';
 import 'package:hotel/services/repos.dart';
 import '../data.dart';