+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-# Change log GrowERP Ecommerce
+# Change log GrowERP Hotel
 
 * 2020/08/07    Initial release
+* 2020/08/07    Mrege from master